[WebXR] Provide a way to bind and unbind IOSurfaces to ANGLE Pbuffers https://bugs.webkit.org/show_bug.cgi?id=226423 <rdar://problem/78652783>
Reviewed by Sam Weinig.

WebXR on Cocoa platforms gets framebuffer texture targets via
IOSurfaces. Those must be attached to ANGLE Pbuffers, which are
then bound to GL textures. There is already some code in
GraphicsContextGL (and in the CoreVideo classes) to do this, but
this is an attempt to make a generic function that will handle all
cases.

* platform/graphics/cocoa/GraphicsContextGLOpenGLCocoa.mm:
(WebCore::GraphicsContextGLOpenGL::createPbufferAndAttachIOSurface):
(WebCore::GraphicsContextGLOpenGL::destroyPbufferAndDetachIOSurface):
* platform/graphics/opengl/GraphicsContextGLOpenGL.h:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@278260 268f45cc-cd09-0410-ab3c-d52691b4dbfc