ANGLE Metal translator pre-rotation code is unused and it asserts when used https://bugs.webkit.org/show_bug.cgi?id=226262
Patch by Kimmo Kinnunen <kkinnunen@apple.com> on 2021-06-01
Reviewed by Dean Jackson.

Remove the pre-rotation code. It is Android specific and most likely it is useful
only when ANGLE is used as the drawing mechanism for the primary app window picture.
Thus most likely it is not useful in WebKit on Cocoa platforms where the window
server does the compositing.

* src/compiler/translator/TranslatorMetalDirect.cpp:
(sh::TranslatorMetalDirect::translateImpl):
* src/compiler/translator/TranslatorMetalDirect.h:
* src/libANGLE/renderer/metal/ContextMtl.h:
* src/libANGLE/renderer/metal/ContextMtl.mm:
(rx::ContextMtl::handleDirtyDriverUniforms):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@278298 268f45cc-cd09-0410-ab3c-d52691b4dbfc