Incorrect text selection when crossing flex item boundary https://bugs.webkit.org/show_bug.cgi?id=226518 <rdar://59487603>
Reviewed by Simon Fraser.

Source/WebCore:

Let's use the more sophisticated "which child is closer to the hit point" logic in RenderBox for the flex (and grid) content.
The rest of the RenderBlock::positionForPoint is meant to deal with BFC content where block level boxes are positioned in block direction, after each other.

Test: editing/selection/selecting-content-by-overshooting-the-container.html

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::positionForPoint):

LayoutTests:

* editing/selection/selecting-content-by-overshooting-the-container-expected.txt: Added.
* editing/selection/selecting-content-by-overshooting-the-container.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@278411 268f45cc-cd09-0410-ab3c-d52691b4dbfc