REGRESSION (Safari 14): Menu items jump around on codelearn.cat https://bugs.webkit.org/show_bug.cgi?id=226608 <rdar://78636003>
Reviewed by Alan Bujtas.

Source/WebCore:

This site changes overflow style dynamically on a stacking-context layer which
has an existing child compositing layer, and the layer has an offset from its
renderer due to the box-shadow.

When this happens, we need to ensure the child layers get a geometry update,
because the newly created clipping layer changes the offset for the child layers.

Test: compositing/overflow/overflow-change-reposition-descendants.html

* rendering/RenderLayerBacking.cpp:
(WebCore::RenderLayerBacking::updateAfterLayout):
* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::layerStyleChanged):

LayoutTests:

* compositing/overflow/overflow-change-reposition-descendants-expected.html: Added.
* compositing/overflow/overflow-change-reposition-descendants.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@278428 268f45cc-cd09-0410-ab3c-d52691b4dbfc