Flaky crash under UserMediaCaptureManagerProxy::SourceProxy::~SourceProxy() on the bots https://bugs.webkit.org/show_bug.cgi?id=226653
Reviewed by Eric Carlson.

The SourceProxy destructor was taking care of calling invalidate() on the SharedRingBufferStorage
before destroying the CARingBuffer on the main thread, to avoid having SourceProxy::storageChanged()
called in the middle of destruction. However, the background thread may still be running at this
point and may reconstruct the RingBuffer right after the invalidate call, causing us to crash
because storageChanged() still gets called in the middle on destruction.

To address the issue, we now make sure to stop the rendering thread before we proceed with the
destruction and invalidate the SharedRingBufferStorage's storage change handler.

* UIProcess/Cocoa/UserMediaCaptureManagerProxy.cpp:
(WebKit::UserMediaCaptureManagerProxy::SourceProxy::~SourceProxy):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@278500 268f45cc-cd09-0410-ab3c-d52691b4dbfc