[GTK][WPE] Signal "window-object-cleared" not emitted unless frame js context is get before https://bugs.webkit.org/show_bug.cgi?id=221771
Reviewed by Michael Catanzaro.

Source/WebKit:

This is because when FrameLoader::dispatchDidClearWindowObjectInWorld() is called, the JSWindowProxy for the
world hasn't been created yet, so it returns early. But we want to dispatch it always for worlds created by the
API, to ensure that WebKitScriptWorld::window-object-cleared is emmitted.

* WebProcess/InjectedBundle/API/glib/WebKitWebPage.cpp: Implement globalObjectIsAvailableForFrame() to force the
creation of the JavaScript context for existing WebKitScriptWorlds.

Tools:

Update window-object-cleared test to check that the signal is emitted for the WebKitScriptWorld objects created
by the extension.

* TestWebKitAPI/Tests/WebKitGLib/TestWebExtensions.cpp:
(UserMessageTest::sendMessage):
(UserMessageTest::viewUserMessageReceived):
(UserMessageTest::waitUntilViewMessagesReceived):
(UserMessageTest::waitUntilViewMessageReceived):
(testWebExtensionWindowObjectCleared):
(beforeAll):
* TestWebKitAPI/Tests/WebKitGLib/WebExtensionTest.cpp:
(windowObjectCleared):
(isolatedWorldWindowObjectCleared):
(webkit_web_extension_initialize_with_user_data):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@278548 268f45cc-cd09-0410-ab3c-d52691b4dbfc