Crash in InsertParagraphSeparatorCommand::doApply https://bugs.webkit.org/show_bug.cgi?id=226527
Patch by Frédéric Wang <fwang@igalia.com> on 2021-06-08
Reviewed by Ryosuke Niwa.

Source/WebCore:

After r273375 and r278002, it is possible that InsertParagraphSeparatorCommand::doApply is
confused by a display: table start block, leading to nullptr crash in the rest of the
function. This patch just excludes that case in order to work around that kind of issue.

Test: fast/editing/insert-paragraph-with-img-in-table-crash.html

* editing/InsertParagraphSeparatorCommand.cpp:
(WebCore::InsertParagraphSeparatorCommand::doApply):

LayoutTests:

Add regression test.

* fast/editing/insert-paragraph-with-img-in-table-crash-expected.txt: Added.
* fast/editing/insert-paragraph-with-img-in-table-crash.html: Added. Note that not having a
new line at the end of the HTML file is required to make the test crash without the patch.
* fast/editing/insert-paragraph-with-text-in-table-crash-expected.txt: Added.
* fast/editing/insert-paragraph-with-text-in-table-crash.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@278593 268f45cc-cd09-0410-ab3c-d52691b4dbfc