REGRESSION (r256095): Adding a border-radius, border, or box-shadow breaks animations from scale(0) https://bugs.webkit.org/show_bug.cgi?id=218371 <rdar://problem/70906316>
Reviewed by Simon Fraser.

Source/WebCore:

When computing an animation's transform extent, we must account for implicit keyframes.

Test: webanimations/accelerated-transform-animation-to-scale-zero-with-implicit-from-keyframe.html

* animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::computeExtentOfTransformAnimation const):

LayoutTests:

Add a test where we have an animation on an element with a a border and a transform animation to scale(0) where the
first keyframe is implicit. This test would fail prior to this patch.

* webanimations/accelerated-transform-animation-to-scale-zero-with-implicit-from-keyframe-expected.html: Added.
* webanimations/accelerated-transform-animation-to-scale-zero-with-implicit-from-keyframe.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@278610 268f45cc-cd09-0410-ab3c-d52691b4dbfc