Add window.internals.log() https://bugs.webkit.org/show_bug.cgi?id=226403
Reviewed by Tim Horton.

window.internals.log() writes output to stderr using WTFLogAlways.
This may be useful for interleaving some output generated from within a
test with WebCore logging, which also goes to stderr. console.log()
doesn't work for this, since that goes to stdout and is captured as
the test output.

* testing/Internals.cpp:
(WebCore::Internals::log):
* testing/Internals.h:
* testing/Internals.idl:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@278690 268f45cc-cd09-0410-ab3c-d52691b4dbfc