[JSC] Remove useDataICInOptimizingJIT option https://bugs.webkit.org/show_bug.cgi?id=226862
Reviewed by Mark Lam.

Source/JavaScriptCore:

Originally this option is added to fix performance problem when using DataIC in DFG / FTL.
But it turned out that that performance regression was caused due to disablement of InlineAccess when
using DataIC. With the configuration using DataIC and InlineAccess, we do not observe performance regression.
So, let's just drop this option.

* jit/JITCode.h:
(JSC::JITCode::useDataIC):
* runtime/Options.cpp:
(JSC::Options::recomputeDependentOptions):
* runtime/OptionsList.h:

Tools:

* Scripts/run-jsc-stress-tests:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@278699 268f45cc-cd09-0410-ab3c-d52691b4dbfc